move vc metadata parser to a separate module
