+++ sdk/core/src/services/parsers/index.ts
@@ -1,0 +1,1 @@
+export { VcMetadataParserFactory } from './VcMetadataParser'
